Fixing f16/bf16 stuff for Device impl for cuda

+++ src/tensor_ops/utilities/device.rs
@@ -110,10 +110,10 @@ impl Device<f32> for crate::tensor::Cpu {}
 impl Device<f64> for crate::tensor::Cpu {}
 
 #[cfg(all(feature = "cuda", feature = "f16"))]
-impl Device<f16> for crate::tensor::Cuda {}
+impl Device<half::f16> for crate::tensor::Cuda {}
 
 #[cfg(all(feature = "cuda", feature = "bf16"))]
-impl Device<bf16> for crate::tensor::Cuda {}
+impl Device<half::bf16> for crate::tensor::Cuda {}
 
 #[cfg(feature = "cuda")]
 impl Device<f32> for crate::tensor::Cuda {}
